Fixed the get/sets, overrides, and added comments.
The gets and sets in several files were improperly done, so I corrected
them. I also fixed the problem with using the override keyword without the
virtual keyword that was causing the compilation error. Finally, I
commented several of the files that were lacking and added a couple of
shell empty methods.