Added ability to load maps from XML and example.
To do this, I created a new project called CustomLoad that contains the
Map class. This is so that it will build the Map class before trying to
load the XML file that depends on it existing. I also added an image with
two seperate tiles in it, and drew them to the screen as an example.
Finally, I created a SpriteFont and drew some text to the screen in the
HUD class.