Changed the way that maps are created and loaded.
This is a MASSIVE change. I completely overhauled the map loading system
to not be created by us, but instead to be using XTiled and the Tiled
application for creating our maps. This should greatly simplify map
creation and maintenance. This is simply a rough version, but it currently
shows the map which it loads from a Tiled .tmx file. I removed the
CustomLoad project and added the XTiled and XTiledExtensions projects.