Changed the tile drawing to move the rectangle.
The rectangle of tiles to draw now moves along with the character. It is
not the best code, but it is functional for now. I had to edit the XTiled
code to pull this off, so it would interface with our Camera class. This
was the best way I could think of doing this at the moment. Also added the
ability to spawn at a particular location on the map.