Controls enabled/disabled
There is a bool in the ControlActions struct that will be enabled if
controls are allowed.  There are disable and enable functions that can
be called to disable controls from being used (during loading and other
times when it maybe necessary).  The reset fuction does not reset the
boolean since we may not always want it to reset every update, so the
enabling function must be called when you disable it at one time.  Now,
it is enabled on creation, but the level, or pantheon, could call it
once loading is completed.