Shield!!
Okay, press space to activate shield, which you wont get hurt, and press
again to disable shield.  As of now, the shield does not drain over
time.  Damage can be caused by pressing Tab, but it does not do damage
only once.  This should not be a problem when damage is implemented in
another place, the problem is the keyboard button states compared to the
amount of times it updates.