Modified the way that Entity keeps location.
Basically, the bounding box and draw box are now stored relative to a
vector called action point. This is referenced by the property called
Location which can be modified to set a new location for the entity.

To change the action point relative to the rest of the entity... you'll
have to wait, since that's what I'm going to commit next.

Right now, there are new constructors for the entity class that may make
it a bit more intensive to actually make a new entity. Sorry about that.

What this does is allow the class to have a fairly good definition
without having to worry about how it works. I hope this kind of makes
sense. Feel free to ask me any questions you may have.