Made bullets rotate and fixed infinite scoping.
The bullets now only have one sprite (the sprite sheet needs to be changed
to reflect that), and it now rotates exactly to the direction of the shot.
Also, when you press the scope button now, it will not follow the
movements of the cursor. It will only scope in the initial direction.