Proof that origin moves things.
I guess I'll have to show this later... today. Wow. So late.

Anyways, this contains some of the debug graphics that I've added for
analysis. It took me a while, but I think I've finally realized that the
origin is somehow modifying the location to which the sprite is drawn.

Unfortunately, there doesn't seem to be any real corelation for this
modification. Analysis with GIMP shows that the pixel counts are correct
and reasonable except after being rotated around a non-zero origin. Then
there are non-correlated values. I'm not sure where they are coming from,
particularly when the sprite isn't being rotated at all.

In any case, I'll stop writing my novel with one final note. It is
probable that I'm missing something simple, so if you can see it, feel
free to show me and tell me what an idiot I am. Thanks.